MediQueryAI/requirements.txt
# =============================================
# MediQuery AI - Complete Requirements
# =============================================

# 🚀 Core Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# 🎯 Modern Frontend (Streamlit)
streamlit==1.28.0
plotly==5.17.0

# 🤖 AI/ML Core Libraries
transformers==4.35.2
torch==2.1.2
torchvision==0.16.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# 🖼️ Image Processing & Computer Vision
pillow==10.1.0
opencv-python==4.8.1.78

# 📄 Document Processing
PyPDF2==3.0.1
pypdf==3.17.0

# 📊 Data Science & Analytics
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
matplotlib==3.7.2
seaborn==0.13.0

# 🌐 HTTP & API Clients
requests==2.31.0
httpx==0.25.2
aiofiles==23.2.1

# ⚙️ Configuration & Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 📈 Additional Data Visualization
altair==5.1.2
plotly-express==0.4.1

# 🔍 Search & NLP Enhancements
nltk==3.8.1
spacy==3.7.2
# python -m spacy download en_core_web_sm

# 💾 Database & Caching (Optional)
redis==5.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.7

# 🛠️ Development & Testing
pytest==7.4.2
pytest-asyncio==0.21.1
black==23.9.1
flake8==6.1.0
jupyter==1.0.0
ipython==8.15.0

# 🔒 Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# 📦 Packaging & Distribution
setuptools==68.2.2
wheel==0.41.2
